CG-0MP12WI75001L9P4: Migrate all example games to consolidated @core-engine/transcript imports
All example game files now import from the canonical consolidated module
instead of the legacy shim paths. The legacy files remain as backward-
compatible re-exports so existing code continues to work.

Games migrated:
- golf/GameTranscript.ts, golf/scenes/GolfOverlayManager.ts
- feudalisim/GameTranscript.ts, feudalisim/scenes/FeudalismOverlayManager.ts
- sushi-go/GameTranscript.ts, sushi-go/scenes/SushiGoOverlayManager.ts
- lost-cities/GameTranscript.ts, lost-cities/scenes/LostCitiesOverlayManager.ts
- beleaguered-castle/GameTranscript.ts
- the-mind/GameTranscript.ts
- main-street/MainStreetTranscript.ts
