+++ example-games/beleaguered-castle/GameTranscript.ts
@@ -24,9 +24,11 @@ import type {
   BCMove,
 } from './BeleagueredCastleState';
 import { FOUNDATION_COUNT, TABLEAU_COUNT } from './BeleagueredCastleState';
-import { snapshotCard } from '../../src/core-engine/TranscriptTypes';
-import type { CardSnapshot } from '../../src/core-engine/TranscriptTypes';
-import { TranscriptRecorderBase } from '../../src/core-engine/TranscriptRecorder';
+import {
+  snapshotCard,
+  TranscriptRecorderBase,
+  type CardSnapshot,
+} from '../../src/core-engine/transcript';
 
 // Re-export so existing consumers that import from this module still work.
 export { snapshotCard };
